requirements.txt
fastapi
sqlalchemy
psycopg2-binary
pydantic
python-multipart
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]
uvicorn
sqlmodel
python-dotenv
ujson
pydantic_settings
pydantic[email]
redis==5.0.1
